% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modselrLOS-package.R
\docType{package}
\name{modselrLOS-package}
\alias{modselrLOS}
\alias{modselrLOS-package}
\title{modselrLOS: Model selection for r-largest order statistics}
\description{
Two algorithms to select a good r-largest order statistics (rLOS) model with an appropriate r are implemented. The first one is the algorithm Survival. The second is the algorithm r median. The first algorithm selects the best model first, then determines an appropriate r for the fixed model. The second algorithm determines an appropriate r, then selects the best model for the fixed r. The pool of models for rLOS are rK4D, rK3D, rGEV, rGLO, rLD, rGGD, and rGD.
}
\author{
\strong{Maintainer}: Yire Shin \email{shinyire87@gmail.com}

Authors:
\itemize{
  \item Jeong-soo Park \email{jsparkstat@gmail.com} [contributor]
}

}
\keyword{internal}
