\name{print.msrlos}
\alias{print.msrlos}
\title{Print msrlos Results}
\usage{\method{print}{msrlos}(x, ...)}
\arguments{
  \item{x}{An object returned by \code{ms.rlos}.}
  \item{...}{Unused.}
}
\value{The input object, invisibly.}
\description{Prints the selected model and \eqn{r} from the four model-selection procedures.}
