% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.fit.park_15Sep25.R
\name{rgd.fit.park}
\alias{rgd.fit.park}
\title{rgd.fit.park}
\usage{
rgd.fit.park(
  xdat,
  r = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  mulink = identity,
  siglink = identity,
  num_inits = 100,
  muinit = NULL,
  siginit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 2000,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rgd.fit.park
}
