% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgev.fit.park_15Sep25.R
\name{rgev.fit.park}
\alias{rgev.fit.park}
\title{rgev.fit.park}
\usage{
rgev.fit.park(xdat = NULL, r = NULL, num_inits = 20, reltol = 1e-06, show = F)
}
\arguments{
\item{show}{}
}
\description{
rgev.fit.park
}
