% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rggd.fit.park_15Sep25.R
\name{rggd.fit.park}
\alias{rggd.fit.park}
\title{rggd.fit.park}
\usage{
rggd.fit.park(
  xdat,
  r = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  hl = NULL,
  mulink = identity,
  siglink = identity,
  hlink = identity,
  num_inits = 100,
  muinit = NULL,
  siginit = NULL,
  hinit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 3000,
  reltol = 1e-06,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rggd.fit.park
}
