% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglo.fit.park.15Sep25.R
\name{rglo.fit.park}
\alias{rglo.fit.park}
\title{rglo.fit.park}
\usage{
rglo.fit.park(
  xdat,
  r = dim(xdat)[2],
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  shl = NULL,
  mulink = identity,
  siglink = identity,
  shlink = identity,
  num_inits = 100,
  muinit = NULL,
  siginit = NULL,
  shinit = NULL,
  show = FALSE,
  maxit = 2000,
  lower = c(-Inf, 0, -1),
  upper = c(Inf, Inf, 1),
  method = "Nelder-Mead",
  low.xi = -0.7,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rglo.fit.park
}
