% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk3d.fit.park_15Sep25.R
\name{rk3d.fit.park}
\alias{rk3d.fit.park}
\title{rk3d.fit.park}
\usage{
rk3d.fit.park(
  xdat,
  r = NULL,
  h.fix = NULL,
  covpar = FALSE,
  penk = NULL,
  penh = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  shl = NULL,
  hl = NULL,
  mulink = identity,
  siglink = identity,
  shlink = identity,
  hlink = identity,
  num_inits = 50,
  muinit = NULL,
  siginit = NULL,
  shinit = NULL,
  hinit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 2000,
  uph = 1.2,
  reltol = 1e-08,
  low.xi = -0.7,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rk3d.fit.park
}
