% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4d.fit.park.15Sep25.R
\name{rk4d.fit.park}
\alias{rk4d.fit.park}
\title{rk4d.fit.park}
\usage{
rk4d.fit.park(
  xdat,
  r = NULL,
  penk = NULL,
  penh = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  shl = NULL,
  hl = NULL,
  mulink = identity,
  siglink = identity,
  shlink = identity,
  hlink = identity,
  num_inits = 50,
  muinit = NULL,
  siginit = NULL,
  shinit = NULL,
  hinit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 5000,
  uph = 2.2,
  reltol = 1e-06,
  low.xi = -0.7,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rk4d.fit.park
}
