% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rld.fit.park.15Sep25.R
\name{rld.fit.park}
\alias{rld.fit.park}
\title{rld.fit.park}
\usage{
rld.fit.park(
  xdat,
  r = NULL,
  ydat = NULL,
  mul = NULL,
  sigl = NULL,
  mulink = identity,
  siglink = identity,
  num_inits = 100,
  muinit = NULL,
  siginit = NULL,
  show = TRUE,
  method = "Nelder-Mead",
  maxit = 10000,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
rld.fit.park
}
