% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_rmed_open_15Sep25.R
\name{sel.rmod.rmed}
\alias{sel.rmod.rmed}
\title{Algoritm r_Median}
\usage{
sel.rmod.rmed(
  xdat,
  maxr = NULL,
  choose = "median",
  h.fix = NULL,
  num_inits = 20,
  sig.ed = 0.05,
  qq = c(0.98, 0.99, 0.995, 0.998),
  true.para = NULL
)
}
\arguments{
\item{xdat}{Data of matrix with \code{R} number of columns for order statistics and n rows for independent observations.}

\item{maxr}{\code{R} represents the maximum, predetermined number of the top order. Default is that R <- ncol(xdat).}

\item{choose}{'median' or 'average' of some \code{r}s, to determine an appropriate \code{r}. Default is 'median'.}

\item{h.fix}{Use when the user have any prior set for the parameter h.}

\item{num_inits}{Number of trials in finding the MLE using numerical optimization routine.}

\item{sig.ed}{Significance level for the ED test. Default is 0.05.}

\item{qq}{Probability values. These values are used for computing T= 1/(1-qq) year return levels.}

\item{true.para}{Useful for simulation study when the user know the true parameter already.}
}
\value{
{
A list containing the following components.

\item{model names}{Model names in the pool of models.}
\item{best.model}{Name of the best model selected.}
\item{opt_r}{Optimal \code{r} for the selected model, which was obtained by the ED test for the best model.}
\item{rhat}{The \code{r} obtained by median or average of \code{r}s from the ED test for the pool of models.}
\item{theta.best.model}{Parameter estimates of the best model.}
\item{rt}{ Return level estimates corresponding to the input qq.}
}
}
\description{
The algorithm \code{r} median determines an appropriate \code{r}, then selects the best model for the fixed \code{r}.
}
\examples{
\dontrun{
sel.rmod.rmed(data)
}
}
