% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_surv_open_24Sep25.R
\name{sel.rmod.surv}
\alias{sel.rmod.surv}
\title{Algoritm survival}
\usage{
sel.rmod.surv(
  xdat,
  maxr = NULL,
  h.fix = NULL,
  mscrit = "AIC",
  num_inits = 20,
  show = TRUE,
  sig.ed = 0.05,
  qq = c(0.98, 0.99, 0.995, 0.998),
  true.para = NULL
)
}
\arguments{
\item{xdat}{Data of matrix with \code{R} number of columns for order statistics and n rows for independent observations.}

\item{maxr}{\code{R} represents the maximum, predetermined number of the top order. Default is that R <- ncol(xdat).}

\item{h.fix}{Use when the user have any prior set for the parameter h.}

\item{mscrit}{Criterion for selecting a model based on the likelihood. Default is "AIC". There are options for "AICc" and "BIC".}

\item{num_inits}{Number of trials in finding the MLE using numerical optimization routine.}

\item{show}{Logical. If it is TRUE, it shows the progress of eliminating the models.}

\item{sig.ed}{Significance level for the ED test. Default is 0.05.}

\item{qq}{Probability values. These values are used for computing T= 1/(1-qq) year return levels.}

\item{true.para}{Useful for simulation study when the user know the true parameter already.}
}
\value{
{
A list containing the following components.

\item{model names}{Model names in the pool of models.}
\item{surv.model}{Name of the best model selected.}
\item{opt_r}{Optimal \code{r} for the selected model, which was obtained by the ED test for the best model.}
\item{theta.best.model}{Parameter estimates of the best model.}
\item{rt}{ Return level estimates corresponding to the input qq.}
}
}
\description{
The algorithm Survival selects the best model first, then determines an appropriate \code{r} for the fixed model.
}
\examples{
\dontrun{
sel.rmod.surv(data)
}

}
